Add tests for C preprocessor parser
I want to extend HIndent's parser for C preprocessing
directives. Before making changes, protect against
regressions by unit testing the existing parser. Expose the
parser through the HIndent.CodeBlock module so tests can
access it directly.

This commit should not change behavior.
